client side lib usage explained, binding and using data from appsettings.json covered.

+++ BLL/Services/CategoryService.cs
@@ -1,6 +1,7 @@
 using BLL.DAL;
 using BLL.Models;
 using BLL.Services.Bases;
+using Microsoft.EntityFrameworkCore;
 
 namespace BLL.Services;
 
@@ -22,6 +23,8 @@ public class CategoryService: Service, ICategoryService
     {
         if (_db.categories.Any(c => c.name == category.name))
             return Error("Category already exists");
+        category.name = category.name?.Trim();
+        category.description = category.description?.Trim();
         _db.categories.Add(category);
         _db.SaveChanges();
         return Success();
@@ -29,14 +32,23 @@ public class CategoryService: Service, ICategoryService
 
     public Service Update(category category)
     {
-        _db.categories.Update(category);
-        _db.SaveChanges();
+        if (_db.categories.Any(c => c.id != category.id && c.name == category.name))
+            return Error("Category already exists");
+        
+        var entity = _db.categories.Find(category.id);
+        entity.name = category.name?.Trim();
+        entity.description = category.description?.Trim();
+        _db.categories.Update(entity);
         return Success();
     }
 
     public Service Delete(int id)
     {
-        var category = _db.categories.Find(id);
+        var category = _db.categories.Include(c => c.products).SingleOrDefault(c => c.id == id);
+
+        if (category is not null && category.products.Any())
+            return Error("Category cannot be delete, it has relational products");
+        
         _db.categories.Remove(category);
         _db.SaveChanges();
         return Success();
